+++ R/cdm.R
@@ -633,9 +633,9 @@ NULL
 #'
 #' @param cdm A cdm object
 #'
-#' @return A list of attributes about the cdm including selected fields from the
-#'   cdm_source table and record counts from the person and observation_period
-#'   tables
+#' @return A named list of attributes about the cdm including selected fields
+#' from the cdm_source table and record counts from the person and
+#' observation_period tables
 #' @export
 #'
 #' @examples
@@ -648,11 +648,8 @@ NULL
 #' DBI::dbDisconnect(con, shutdown = TRUE)
 #' }
 snapshot <- function(cdm) {
-  assert_tables(cdm,
-                tables = c("cdm_source",
-                           "person",
-                           "observation_period",
-                           "vocabulary"))
+  assert_tables(cdm, tables = c("cdm_source", "vocabulary"), empty.ok = TRUE)
+  assert_tables(cdm, tables = c("person", "observation_period"))
 
   person_cnt <- dplyr::tally(cdm$person, name = "n") %>% dplyr::pull(.data$n)
 
@@ -664,9 +661,22 @@ snapshot <- function(cdm) {
     dplyr::filter(.data$vocabulary_id == "None") %>%
     dplyr::pull(.data$vocabulary_version)
 
+  if (length(vocab_version) == 0) {
+    vocab_version <- NA_character_
+  }
+
   cdm_source_name <- cdm$cdm_source %>% dplyr::pull(.data$cdm_source_name)
 
-  dplyr::collect(cdm$cdm_source) %>%
+  cdm_source <- dplyr::collect(cdm$cdm_source)
+  if (nrow(cdm_source) == 0) {
+    cdm_source <- dplyr::tibble(vocabulary_version = vocab_version,
+                                cdm_source_name = "",
+                                cdm_holder = "",
+                                cdm_release_date = "",
+                                cdm_version = attr(cdm, "cdm_version"))
+  }
+
+  cdm_source %>%
     dplyr::mutate(vocabulary_version = dplyr::coalesce(.env$vocab_version,
                                                        .data$vocabulary_version)) %>%
     dplyr::mutate(
@@ -683,9 +693,14 @@ snapshot <- function(cdm) {
       "observation_period_cnt"
     ) %>%
     as.list() %>%
+    c(list(cdm_schema = attr(cdm, "cdm_schema"),
+           write_schema = attr(cdm, "write_schema"),
+           cdm_name = attr(cdm, "cdm_name"))) %>%
     magrittr::set_class("cdm_snapshot")
+
 }
 
+#' @export
 print.cdm_snapshot <- function(x, ...) {
   cli::cat_rule(x$cdm_source_name)
   purrr::walk2(names(x[-1]), x[-1], ~ cli::cat_bullet(.x, ": ", .y))